Update Bundler version for Ruby 3.2
Previously, Bundler would use version 1.14.6 specified in the
Gemfile.lock. However, this version of Bundler uses the #untaint method
which was removed in Ruby 3.2.

To fix this, the locked version of Bundler has been updated. The version
chosen is the last 2.3.x release as 2.4 drops support for Rubyies 2.3 -
2.5. If dropping support for EoL Rubies is not an issue then the Bundler
version could be upgraded further.