Build into template directories instead of /out
The spec isn't clear about where templates should be built, but this is
the pattern that most template repositories end up using. Additionally,
building into /out causes templates with the same output directory and
extension to conflict.

Building into /out was previously introduced in response to an issue
describing the previous behavior as non-standard:

```
/templates/out
```

vs

```
/templates
/out
```

It seems like `/templates` was misunderstood as the builder's directory
of all templates, but I think the issue creator was referring to the
directory of templates inside a templare repository. The original code
would build inside template repositories' `/templates` folder, but they
should have been building at the repository root (which this commit
does).